+++ python/bindings/openravepy_kinbody.cpp
@@ -57,6 +57,26 @@ KinBody::KinBodyInfoPtr ExtractKinBodyInfo(object obj)
     return NULL;
 }
 
+void ExtractGeometryInfoArray(py::object pyGeometryInfoList, std::vector<KinBody::GeometryInfo>& vGeometryInfos)
+{
+    vGeometryInfos.clear();
+    if( IS_PYTHONOBJECT_NONE(pyGeometryInfoList) ) {
+        return;
+    }
+    const size_t arraySize = len(pyGeometryInfoList);
+    vGeometryInfos.resize(arraySize);
+
+    for(size_t igeom = 0; igeom < arraySize; igeom++) {
+        extract_<OPENRAVE_SHARED_PTR<PyGeometryInfo> > pyGeometryInfo(pyGeometryInfoList[py::to_object(igeom)]);
+        if (pyGeometryInfo.check()) {
+            ((OPENRAVE_SHARED_PTR<PyGeometryInfo>)pyGeometryInfo)->FillGeometryInfo(vGeometryInfos[igeom]);
+        }
+        else{
+            throw openrave_exception("Bad GeometryInfo");
+        }
+    }
+}
+
 std::vector<KinBody::LinkInfoPtr> ExtractLinkInfoArray(object pyLinkInfoList)
 {
     if( IS_PYTHONOBJECT_NONE(pyLinkInfoList) ) {
@@ -77,8 +97,8 @@ std::vector<KinBody::LinkInfoPtr> ExtractLinkInfoArray(object pyLinkInfoList)
             }
         }
     }
-    catch(...) {
-        RAVELOG_WARN("Cannot do ExtractArray for LinkInfos");
+    catch(std::exception& ex) {
+        RAVELOG_WARN_FORMAT("Cannot do ExtractArray for LinkInfos: %s", ex.what());
     }
     return vLinkInfos;
 }
@@ -335,9 +355,15 @@ void PyGeometryInfo::DeserializeJSON(object obj, dReal fUnitScale, object option
     Init(*pgeominfo);
 }
 
-KinBody::GeometryInfoPtr PyGeometryInfo::GetGeometryInfo() {
+KinBody::GeometryInfoPtr PyGeometryInfo::GetGeometryInfo()
+{
     KinBody::GeometryInfoPtr pinfo(new KinBody::GeometryInfo());
-    KinBody::GeometryInfo& info = *pinfo;
+    FillGeometryInfo(*pinfo);
+    return pinfo;
+}
+
+void PyGeometryInfo::FillGeometryInfo(KinBody::GeometryInfo& info)
+{
     info.SetTransform(ExtractTransform(_t));
     info._vGeomData = ExtractVector<dReal>(_vGeomData);
     info._vGeomData2 = ExtractVector<dReal>(_vGeomData2);
@@ -413,7 +439,36 @@ KinBody::GeometryInfoPtr PyGeometryInfo::GetGeometryInfo() {
 #endif
 
     }
-    return pinfo;
+}
+
+object PyGeometryInfo::GetBoxHalfExtents()
+{
+    return toPyVector3(ExtractVector<dReal>(_vGeomData));
+}
+
+object PyGeometryInfo::GetCageBaseHalfExtents()
+{
+    return toPyVector3(ExtractVector<dReal>(_vGeomData));
+}
+
+object PyGeometryInfo::GetContainerOuterExtents()
+{
+    return toPyVector3(ExtractVector<dReal>(_vGeomData));
+}
+
+object PyGeometryInfo::GetContainerInnerExtents()
+{
+    return toPyVector3(ExtractVector<dReal>(_vGeomData2));
+}
+
+void PyGeometryInfo::SetContainerOuterExtents(object oOuterExtents)
+{
+    _vGeomData = oOuterExtents;
+}
+
+void PyGeometryInfo::SetContainerInnerExtents(object oInnerExtents)
+{
+    _vGeomData2 = oInnerExtents;
 }
 
 PyLinkInfo::PyLinkInfo() {
@@ -4858,6 +4913,12 @@ void KinBodyInitializer::init_openravepy_kinbody()
                           .def_readwrite("_vPositiveCropContainerEmptyMargins", &PyGeometryInfo::_vPositiveCropContainerEmptyMargins)
                           .def("ComputeInnerEmptyVolume",&PyGeometryInfo::ComputeInnerEmptyVolume, DOXY_FN(GeomeryInfo,ComputeInnerEmptyVolume))
                           .def("ComputeAABB",&PyGeometryInfo::ComputeAABB, PY_ARGS("transform") DOXY_FN(GeomeryInfo,ComputeAABB))
+                          .def("GetBoxHalfExtents",&PyGeometryInfo::GetBoxHalfExtents, DOXY_FN(GeomeryInfo,GetBoxHalfExtents))
+                          .def("GetCageBaseHalfExtents",&PyGeometryInfo::GetCageBaseHalfExtents, DOXY_FN(GeomeryInfo,GetCageBaseHalfExtents))
+                          .def("GetContainerOuterExtents",&PyGeometryInfo::GetContainerOuterExtents, DOXY_FN(GeomeryInfo,GetContainerOuterExtents))
+                          .def("GetContainerInnerExtents",&PyGeometryInfo::GetContainerInnerExtents, DOXY_FN(GeomeryInfo,GetContainerInnerExtents))
+                          .def("SetContainerOuterExtents",&PyGeometryInfo::SetContainerOuterExtents, PY_ARGS("outerExtents") DOXY_FN(GeomeryInfo,GetContainerInnerExtents))
+                          .def("SetContainerInnerExtents",&PyGeometryInfo::SetContainerInnerExtents, PY_ARGS("innerExtents") DOXY_FN(GeomeryInfo,GetContainerInnerExtents))
 #ifdef USE_PYBIND11_PYTHON_BINDINGS
                           .def("SerializeJSON", &PyGeometryInfo::SerializeJSON,
                                "unitScale"_a = 1.0,